test: merge duplicated component spec pairs
Nine spec basenames existed in both tests/unit/components and
tests/component, testing the same source component. Kept the larger
file in each pair, folded in unique non-vacuous tests from the
smaller file, dropped genuinely vacuous or duplicate ones, and
removed the smaller file.

Also fixes a branch-coverage regression the merge introduced in
DataSourceRunTable.vue's showKey computation, caught by comparing
coverage-summary.json before/after over tests/component plus
tests/unit/components in single invocations.
