+++ tests/unit/components/cohort-builder/CriteriaEventCard.spec.ts
@@ -5,7 +5,11 @@ import { createPinia, setActivePinia } from 'pinia'
 import * as components from 'vuetify/components'
 import * as directives from 'vuetify/directives'
 import CriteriaEventCard from '@/components/cohort-builder/CriteriaEventCard.vue'
-import type { CohortEvent } from '@/models/cohort.types'
+import {
+  CriteriaSelectionKey,
+  type CriteriaSelectionService,
+} from '@/composables/useCriteriaSelection'
+import type { CohortEvent, ConceptSetReference } from '@/models/cohort.types'
 
 vi.mock('@/composables/useI18n', async () => {
   const { mockUseI18n } = await import('../../../helpers/i18n-mock')
@@ -399,5 +403,80 @@ describe('CriteriaEventCard', () => {
       const wrapper = mountComponent(event)
       expect(wrapper.find('.event-type-label').exists()).toBe(true)
     })
+  })
+})
+
+describe('CriteriaEventCard, concept-set selection', () => {
+  const createWrapperWithSelection = (
+    event: CohortEvent,
+    selection?: CriteriaSelectionService
+  ) => {
+    const pinia = createPinia()
+    setActivePinia(pinia)
+    return mount(CriteriaEventCard, {
+      global: {
+        plugins: [vuetify, pinia],
+        provide: selection ? { [CriteriaSelectionKey as symbol]: selection } : {},
+      },
+      props: { event, section: 'criteriaGroup' },
+    })
+  }
+
+  it('requests the concept set via the selection service and applies the result', async () => {
+    let deliver: ((cs: ConceptSetReference) => void) | undefined
+    const selection: CriteriaSelectionService = {
+      requestConceptSet: vi.fn(cb => {
+        deliver = cb
+      }),
+      requestConcepts: vi.fn(),
+      editConceptSet: vi.fn(),
+    }
+    const wrapper = createWrapperWithSelection(createMockEvent(), selection)
+
+    await wrapper.find('[data-testid="concept-set-picker"]').trigger('click')
+
+    expect(selection.requestConceptSet).toHaveBeenCalledTimes(1)
+    expect(wrapper.emitted('select-concept-set')).toBeFalsy()
+
+    // The owner delivers the picked set; the card applies it through its
+    // ordinary update emit, which is what makes selection depth-independent.
+    const picked: ConceptSetReference = { id: 4, name: 'Diabetes', items: [] }
+    deliver?.(picked)
+    const updates = wrapper.emitted('update')
+    expect(updates).toBeTruthy()
+    expect((updates![0]![0] as CohortEvent).conceptSet).toEqual(picked)
+  })
+
+  it('falls back to the legacy select-concept-set emit without a service', async () => {
+    const wrapper = createWrapperWithSelection(createMockEvent())
+
+    await wrapper.find('[data-testid="concept-set-picker"]').trigger('click')
+
+    expect(wrapper.emitted('select-concept-set')).toBeTruthy()
+  })
+
+  it('opens the editor via the selection service when the chip is clicked', async () => {
+    const selection: CriteriaSelectionService = {
+      requestConceptSet: vi.fn(),
+      requestConcepts: vi.fn(),
+      editConceptSet: vi.fn(),
+    }
+    const conceptSet = { id: 9, name: 'Hypertension', items: [] }
+    const wrapper = createWrapperWithSelection(createMockEvent({ conceptSet }), selection)
+
+    await wrapper.find('[data-testid="selected-concept-set"]').trigger('click')
+
+    expect(selection.editConceptSet).toHaveBeenCalledWith(conceptSet)
+    expect(wrapper.emitted('edit-concept-set')).toBeFalsy()
+  })
+
+  it('falls back to the legacy edit-concept-set emit without a service', async () => {
+    const conceptSet = { id: 9, name: 'Hypertension', items: [] }
+    const wrapper = createWrapperWithSelection(createMockEvent({ conceptSet }))
+
+    await wrapper.find('[data-testid="selected-concept-set"]').trigger('click')
+
+    expect(wrapper.emitted('edit-concept-set')).toBeTruthy()
+    expect(wrapper.emitted('edit-concept-set')![0]).toEqual([conceptSet])
   })
 })
